+++ src/flask_exts/templating/__init__.py
@@ -1,20 +1,19 @@
+import os.path as op
 from flask import Blueprint
 from ..utils.form import is_hidden_field, get_table_titles
 from ..utils.csrf import generate_csrf
+from .theme import DefaultTheme
 
 
 def template_init_app(app):
 
-    if not hasattr(app, "extensions"):
-        app.extensions = {}
-    app.extensions["templating"] = "none"
-
     blueprint = Blueprint(
-        "templating",
+        "template",
         __name__,
-        template_folder="../templates",
-        static_url_path='/templating/static',
-        static_folder="../static",
+        url_prefix="/template",
+        template_folder=op.join("..", "templates"),
+        static_folder=op.join("..", "static"),
+        # static_url_path='/template/static',
     )
     app.register_blueprint(blueprint)
 
@@ -22,10 +21,20 @@ def template_init_app(app):
     app.jinja_env.globals["is_hidden_field"] = is_hidden_field
     app.jinja_env.globals["get_table_titles"] = get_table_titles
 
-    template_name = app.config.get("TEMPLATE_NAME")
+    if app.config.get("TEMPLATE_THEME"):
+        theme = app.config.get("TEMPLATE_THEME")
+    else:
+        theme = DefaultTheme()
 
-    if template_name == "bootstrap":
-        from .bootstrap import Bootstrap
+    app.extensions["template"] = theme
 
-        bootstrap = Bootstrap()
-        bootstrap.init_app(app)
+    app.jinja_env.globals["theme"] = theme
+
+    # app.config.setdefault("BOOTSTRAP_TABLE_VIEW_TITLE", "View")
+    # app.config.setdefault("BOOTSTRAP_TABLE_EDIT_TITLE", "Edit")
+    # app.config.setdefault("BOOTSTRAP_TABLE_DELETE_TITLE", "Delete")
+    # app.config.setdefault("BOOTSTRAP_TABLE_NEW_TITLE", "New")
+
+    # @app.context_processor
+    # def get_theme():
+    #     return {"theme": theme}