+++ tests/templating/bootstrap4/test_render_icon.py
@@ -91,23 +91,3 @@ def test_render_icon(app, client):
     data = response.get_data(as_text=True)
     assert "bootstrap-icons.svg#heart" in data
     assert "<desc>A heart.</desc>" in data
-
-
-def test_render_icon_config(app, client):
-    app.config["BOOTSTRAP_ICON_SIZE"] = 100
-    app.config["BOOTSTRAP_ICON_COLOR"] = "success"
-
-    @app.route("/icon")
-    def icon():
-        return render_template_string(
-            """
-            {% from 'macro/icon.html' import render_icon %}
-                {{ render_icon('heart') }}
-            """
-        )
-
-    response = client.get("/icon")
-    data = response.get_data(as_text=True)
-    assert 'width="100"' in data
-    assert 'height="100"' in data
-    assert "text-success" in data