+++ CHANGES.rst
@@ -1,3 +1,11 @@
+v0.1.7
+------
+
+Released 2024-10-13
+
+- add theme for templating
+- admin view framework
+
 v0.1.6
 ------
 
@@ -5,7 +13,7 @@ Released 2024-09-14
 
 - merge translations of wtforms, flask_exts and app into flask_babel.get_translations()
 - update FileField and ImageField
-- add url for templating.static
+- add url for template.static
 - use namespace ns.bootstrap to control bootstrap's css and js
 
 v0.1.5